// src/api/generated/model/expenseResponse.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * TaxMate API
 * 프리랜서 세금 관리 서비스 API
 * OpenAPI spec version: 1.0.0
 */
import type { ExpenseResponseCategory } from './expenseResponseCategory';

export interface ExpenseResponse {
  id: string;
  title: string;
  amount: number;
  expenseDate: string;
  category: ExpenseResponseCategory;
  categoryDisplayName: string;
  memo?: string;
  isDeductible: boolean;
  createdAt?: string;
}